staging_meetup_extract/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0

